requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
sentence-transformers==3.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Multi-Agent System (for future LangGraph integration)
# Uncomment when ready to add LangGraph workflow:
# langgraph==0.0.40
# langchain==0.1.0
# langchain-community==0.0.20

# RAG Pipeline Dependencies (for future vector DB integration)
# Uncomment when ready for advanced RAG:
# chromadb==0.4.22
# openai==1.12.0